Refactor : api & db 위치 변경
