+++ src/pages/playlist/hooks/useUpdatePlaylist.ts
@@ -1,6 +1,6 @@
 import { useMutation } from "@tanstack/react-query";
-import { patchPlaylist } from "@/db/playlist";
-import { createVideo, deleteVideosByIds } from "@/db/video";
+import { patchPlaylist } from "@/shared/api/playlist";
+import { createVideo, deleteVideosByIds } from "@/shared/api/video";
 import { Video } from "@/types/video";
 import { diffVideoList } from "@/pages/playlist/utils/diffVideoList";
