+++ src/pages/profile/hooks/useOtherUser.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
-import { getUserByRandomId } from "@/db/users";
-import { User } from "@/db/users";
-import { QUERY_KEYS } from "@/constants/queryKey";
+import { getUserByRandomId } from "@/shared/api/users";
+import { User } from "@/shared/api/users";
+import { QUERY_KEYS } from "@/shared/constants/queryKey";
 
 const useOtherUser = (randomId: number) => {
   const { data, isLoading, isError } = useQuery<User | null>({
