+++ src/pages/storage/hooks/useFollowStatus.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { getFollowStatus, postFollow, deleteFollow } from "@/db/follow";
+import { getFollowStatus, postFollow, deleteFollow } from "@/shared/api/follow";
 import { useUserStore } from "@/stores/userStore";
-import { QUERY_KEYS } from "@/constants/queryKey";
+import { QUERY_KEYS } from "@/shared/constants/queryKey";
 
 const useFollowStatus = (targetId?: number) => {
   const queryClient = useQueryClient();
